back/requirements.txt
# -------------------------------------------------------------------------------------------------
# Base
# -------------------------------------------------------------------------------------------------
psycopg2-binary                 == 2.9.2    # Python-PostgreSQL Database Adapter
PyYAML                          == 6.0.1    # PyYAML is a YAML parser and emitter for Python.
# -------------------------------------------------------------------------------------------------
# Django
# -------------------------------------------------------------------------------------------------
django                          == 4.2.2    #
django-cors-headers             == 4.2.0    # for handling the server headers required for Cross-Origin Resource Sharing (CORS).
django-debug-toolbar            == 3.4.0    # A configurable set of panels that display various debug information about the current request/response.
django-extensions               == 3.2.1    # Django Extensions is a collection of custom extensions for the Django Framework.
djangorestframework             == 3.14.0   # Web APIs for Django, made easy.
drf-flex-fields                 == 1.0.0    # Flexible, dynamic fields and nested resources for Django REST Framework serializers.
# -------------------------------------------------------------------------------------------------
# tools for testing
# -------------------------------------------------------------------------------------------------
pytest                  == 7.2.0    # The framework makes it easy to write small tests, yet scales to support complex functional testing for applications and libraries.
pytest-mock             == 3.10.0   # Thin-wrapper around the mock package for easier use with pytest
pytest-django           == 4.5.2    # A plugin for pytest that provides a set of useful tools for testing Django applications and projects
pytest-cov              == 4.0.0    # This plugin produces coverage reports.
pytest-xdist[psutil]    == 3.2.0    # Run test in parallel.
pylint                  == 2.15.7   # A static code analyser for Python.
pylint-django           == 2.5.3    # A Pylint plugin for improving code analysis when analysing code using Django.
